syslog-ng-common: rename option_error module to error
