+++ syslog-ng-rs/syslog-ng-common/src/proxies/parser/error.rs
